moving problem test to separate project
